backend/requirements.txt
functions-framework==3.*
google-cloud-aiplatform
Pillow
numpy
